+++ cmd/deletion/delete.go
@@ -8,9 +8,10 @@ func NewDeleteCmd() *cobra.Command {
 
 	var cmdDelete = &cobra.Command{
 		Use:   "delete",
-		Short: "delete topics",
+		Short: "delete topics, acls",
 	}
 
 	cmdDelete.AddCommand(newDeleteTopicCmd())
+	cmdDelete.AddCommand(newDeleteAclCmd())
 	return cmdDelete
 }