+++ cmd/get/get.go
@@ -9,11 +9,12 @@ func NewGetCmd() *cobra.Command {
 	var cmdGet = &cobra.Command{
 		Use:     "get",
 		Aliases: []string{"list"},
-		Short:   "get info about topics, consumerGroups",
+		Short:   "get info about topics, consumerGroups, acls",
 	}
 
 	cmdGet.AddCommand(newGetTopicsCmd())
 	cmdGet.AddCommand(newGetConsumerGroupsCmd())
+	cmdGet.AddCommand(newGetAclCmd())
 
 	return cmdGet
 }