+++ test_util/test_util.go
@@ -65,11 +65,23 @@ func StartUnitTest(t *testing.T) {
 }
 
 func StartIntegrationTest(t *testing.T) {
+	StartIntegrationTestWithContext(t, "default")
+}
+
+func StartIntegrationTestWithContext(t *testing.T, context string) {
 	if testing.Short() {
 		t.Skip("skipping integration test")
 	}
 
+	SwitchContext(context)
+
 	startTest(t, "integration-test.log")
+}
+
+func SwitchContext(context string) {
+	if err := os.Setenv("CURRENT_CONTEXT", context); err != nil {
+		panic(err)
+	}
 }
 
 func startTest(t *testing.T, logFilename string) {
@@ -106,6 +118,11 @@ func AssertEquals(t *testing.T, expected, actual string) {
 }
 
 func AssertErrorContains(t *testing.T, expected string, err error) {
+
+	if err == nil {
+		t.Fatalf("expected error to contain: %s\n: %v", expected, "nil")
+	}
+
 	if !strings.Contains(err.Error(), expected) {
 		t.Fatalf("expected error to contain: %s\n: %v", expected, err)
 	}
@@ -129,7 +146,7 @@ func GetPrefixedName(prefix string) string {
 
 func WithoutBrokerReferences(output string) string {
 
-	brokerAddressRegex := regexp.MustCompile(`localhost:\d9092`)
+	brokerAddressRegex := regexp.MustCompile(`localhost:\d909[2|3]`)
 	withoutBrokerAddresses := brokerAddressRegex.ReplaceAllString(output, "any-broker")
 
 	brokerIdRegex := regexp.MustCompile(`([^\d\w])(101|102|103)([^\d\w])`)